+++ app/templates/views/dashboard.html
@@ -11,14 +11,14 @@ GOV.UK Notify | Dashboard
 		<h1 class="heading-xlarge">Dashboard</h1>
 
 		<ul>
-			<li><a href="send-sms">Send text messages</a></li>
-			<li><a href="send-email">Send email messages</a></li>
-			<li><a href="jobs">View notifications activity</a></li>
-			<li><a href="user-profile">User profile</a></li>
-			<li><a href="manage-users">Manage users</a></li>
-			<li><a href="manage-templates">Manage templates</a></li>
-			<li><a href="service-settings">Service settings</a></li>
-			<li><a href="api-keys">API keys and documentation</a></li>
+			<li><a href="{{ url_for('.sendsms') }}">Send text messages</a></li>
+			<li><a href="{{ url_for('.sendemail') }}">Send email messages</a></li>
+			<li><a href="{{ url_for('.showjobs') }}">View notifications activity</a></li>
+			<li><a href="{{ url_for('.userprofile') }}">User profile</a></li>
+			<li><a href="{{ url_for('.manageusers') }}">Manage users</a></li>
+			<li><a href="{{ url_for('.managetemplates')}}">Manage templates</a></li>
+			<li><a href="{{ url_for('.servicesettings') }}">Service settings</a></li>
+			<li><a href="{{ url_for('.apikeys') }}">API keys and documentation</a></li>
   		</ul>
 	</div>
 </div>
